Fix streaming output corruption with copilot (#1261)
* Shortcut other steps if one holds the output

We had a bug in the output pipeline where if one step returned []
meaning that the output chunk should be held off, all the current chunks
would continue to run and potentially modify the context.

* Shortcut buffered PII sooner, as soon as the buffer can't be a UUID

Our PII refaction format is `#UUID#`. Our code was finding an opening #,
then checking for a closing matching # or end of the output. For
copilot, however, this meant that we were buffering the whole file,
because the filename comes in this format:

```
```

This means we would keep searching for the closing hash which never
came. Instead, buffer only as long as the context between the hashes can
reasonably be a UUID.

Fixes: #1250

---------

Co-authored-by: Luke Hinds <luke@stacklok.com>